add integration test for native-tls module
