+++ native-tls/src/lib.rs
@@ -24,9 +24,9 @@ extern crate tokio_core;
 extern crate tokio_tls;
 extern crate trust_dns;
 
-mod tls_client_connection;
-mod tls_client_stream;
-mod tls_stream;
+pub mod tls_client_connection;
+pub mod tls_client_stream;
+pub mod tls_stream;
 
 pub use self::tls_client_connection::{TlsClientConnection, TlsClientConnectionBuilder};
 pub use self::tls_client_stream::{TlsClientStream, TlsClientStreamBuilder};
