--- server/tests/named_test_configs/sec/.gitignore
@@ -1,1 +1,0 @@
-*
